requirements.txt
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
Flask-Marshmallow==0.14.0
Flask-JWT-Extended==4.2.3
Werkzeug==1.0.1
SQLAlchemy==1.3.24
psycopg2-binary==2.9.3
marshmallow-sqlalchemy==0.26.1
python-dotenv==0.19.2
email-validator==1.1.3
MarkupSafe==2.0.1
Jinja2==2.11.3
